Generalize offset calculation to all voices with simultaneous notes
