+++ src/MusicalScore/Graphical/GraphicalNote.ts
@@ -38,6 +38,8 @@ export class GraphicalNote extends GraphicalObject {
     public numberOfDots: number;
     public rules: EngravingRules;
     public staffLine: number;
+    public baseFingeringXOffset: number;
+    public baseStringNumberXOffset: number;
 
     public Transpose(keyInstruction: KeyInstruction, activeClef: ClefInstruction, halfTones: number, octaveEnum: OctaveEnum): Pitch {
         let transposedPitch: Pitch = this.sourceNote.Pitch;
